pyproject.toml
[project]
name = "pygame-library"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.7",
]

[dependency-groups]
dev = [
    "pygbag>=0.9.3",
]
